fix: .NET extension unit tests
